Introduce uchiwa users in ansible-uchiwa
Uchiwa users provides a way to specify a list of user credentials to
give access to the uchiwa dashboard, rather than being restricted to one
only. Users takes precedence over 'user' and 'password'. If not
specified 'user' and 'password' will be used instead.

This new feature was included in uchiwa 0.10.0 and now is supported by
this ansible role.